+++ Sources/Core/Presenter/Channels/ChannelsPresenter+Rx.swift
@@ -43,6 +43,17 @@ public extension Reactive where Base == ChannelsPresenter {
             .do(onNext: { [weak base] _ in base?.rx.removeFromItems(channelPresenter) })
             .asDriver(onErrorJustReturn: .empty)
     }
+    
+    /// Delete a channel and remove a channel presenter from items.
+    ///
+    /// - Parameters:
+    ///   - channelPresenter: a channel presenter.
+    func delete(_ channelPresenter: ChannelPresenter) -> Driver<Channel> {
+        channelPresenter.channel.rx
+            .delete()
+            .do(onNext: { [weak base] _ in base?.rx.removeFromItems(channelPresenter) })
+            .asDriver(onErrorJustReturn: .unused)
+    }
 }
 
 private extension Reactive where Base == ChannelsPresenter {